Removing lowlevel dependencies from low_level_perception
