+++ droidlet/lowlevel/minecraft/mc_util.py
@@ -98,17 +98,6 @@ def cluster_areas(areas):
     return clustered_areas
 
 
-def diag_adjacent(p):
-    """Return the adjacent positions to p including diagonal adjaceny"""
-    return [
-        (x, y, z)
-        for x in range(p[0] - 1, p[0] + 2)
-        for y in range(p[1] - 1, p[1] + 2)
-        for z in range(p[2] - 1, p[2] + 2)
-        if (x, y, z) != p
-    ]
-
-
 def discrete_step_dir(agent):
     """Discretized unit vector in the direction of agent's yaw
