+++ droidlet/perception/craftassist/low_level_perception.py
@@ -1,15 +1,9 @@
 """
 Copyright (c) Facebook, Inc. and its affiliates.
 """
-import os
-import sys
 import numpy as np
-
-from droidlet.lowlevel.minecraft.mc_util import XYZ, IDM, pos_to_np, euclid_dist, diag_adjacent
-from droidlet.base_util import to_block_pos
 from typing import Tuple, List
-from droidlet.lowlevel.minecraft.craftassist_cuberite_utils.block_data import BORING_BLOCKS
-
+from droidlet.base_util import diag_adjacent, to_block_pos, XYZ, IDM, pos_to_np, euclid_dist
 from droidlet.memory.memory_nodes import PlayerNode, AttentionNode
 from droidlet.memory.craftassist.mc_memory_nodes import BlockObjectNode
 
@@ -55,6 +49,7 @@ class LowLevelMCPerception:
         """
         # FIXME (low pri) remove these in code, get from sql
         self.agent.pos = to_block_pos(pos_to_np(self.agent.get_player().pos))
+        boring_blocks = self.agent.low_level_data["boring_blocks"]
 
         if self.agent.count % self.perceive_freq == 0 or force:
             for mob in self.agent.get_mobs():
@@ -84,7 +79,7 @@ class LowLevelMCPerception:
 
         # use safe_get_changed_blocks to deal with pointing
         for (xyz, idm) in self.agent.safe_get_changed_blocks():
-            self.on_block_changed(xyz, idm)
+            self.on_block_changed(xyz, idm, boring_blocks)
 
     def update_self_memory(self):
         """Update agent's current position and attributes in memory"""
@@ -142,12 +137,12 @@ class LowLevelMCPerception:
                 return p
         return None
 
-    def on_block_changed(self, xyz: XYZ, idm: IDM):
+    def on_block_changed(self, xyz: XYZ, idm: IDM, boring_blocks: Tuple[int]):
         """Update the state of the world when a block is changed."""
         # TODO don't need to do this for far away blocks if this is slowing down bot
         self.maybe_remove_inst_seg(xyz)
         self.maybe_remove_block_from_memory(xyz, idm)
-        self.maybe_add_block_to_memory(xyz, idm)
+        self.maybe_add_block_to_memory(xyz, idm, boring_blocks)
 
     def clear_air_surrounded_negatives(self):
         pass
@@ -166,12 +161,12 @@ class LowLevelMCPerception:
 
     # clean all this up...
     # eventually some conditions for not committing air/negative blocks
-    def maybe_add_block_to_memory(self, xyz: XYZ, idm: IDM, agent_placed=False):
+    def maybe_add_block_to_memory(self, xyz: XYZ, idm: IDM, boring_blocks: Tuple[int], agent_placed=False):
         """Update blocks to memory when any change in the environment
         is caused either by agent or player"""
         if not agent_placed:
             interesting, player_placed, agent_placed = self.is_placed_block_interesting(
-                xyz, idm[0]
+                xyz, idm[0], boring_blocks
             )
         else:
             interesting = True
@@ -245,7 +240,7 @@ class LowLevelMCPerception:
                 self.agent.areas_to_perceive.append((xyz, 3))
 
     # FIXME move removal of block to parent
-    def is_placed_block_interesting(self, xyz: XYZ, bid: int) -> Tuple[bool, bool, bool]:
+    def is_placed_block_interesting(self, xyz: XYZ, bid: int, boring_blocks: Tuple[int]) -> Tuple[bool, bool, bool]:
         """Return three values:
         - bool: is the placed block interesting?
         - bool: is it interesting because it was placed by a player?
@@ -266,6 +261,6 @@ class LowLevelMCPerception:
                 interesting = True
                 if not agent_placed:
                     player_placed = True
-        if bid not in BORING_BLOCKS:
+        if bid not in boring_blocks:
             interesting = True
         return interesting, player_placed, agent_placed
